Use new-sides.
